Added   * global UNITTESTING boolean to traitsui/editor.py in order to reraise errors in unittesting.   * show_error_dialog attribute to the EditorFactory class in traitsui/editor_factory.py
Changed Default low and high value attributes of ToolkitEditorFactory from 0 and 1 to -sys.float_info.max and sys.float_info.max, respectively, which avoids the situation that the lower bound is larger than the upper bound and vice versa.

Refactored range_editor.py of qt4 and wx backend so that the codebase and the behavior of the rangeeditors are more similar: Divided the code base into make_controller and do_layout.
Replaced unsafe eval statements with ast.literal_eval and the _validate function.
Added SimpleSliderEditor
Fixed inconsistent LargeRangeSliderEditor
Added step attribute to SimpleSpinEditor as well as the possibility to increase the step with factor 2, 10, or 100 if Shift, Ctrl or Alt are pressed at the same time the spincontroller is increased or decreased.
replaced import "from cgi import escape" to "from html import escape" in wx/ui_panel.py